split test suite for better parallelism
