+++ tests/ImageProcessing/Photo/PhotoCopyTest.php
@@ -16,7 +16,7 @@
  * @noinspection PhpUnhandledExceptionInspection
  */
 
-namespace Tests\Feature_v2\Photo;
+namespace Test\ImageProcessing\Photo;
 
 use Tests\Feature_v2\Base\BaseApiWithDataTest;
