+++ tests/ImageProcessing/Photo/PhotoZipUploadTest.php
@@ -16,7 +16,7 @@
  * @noinspection PhpUnhandledExceptionInspection
  */
 
-namespace Tests\Feature_v2\Photo;
+namespace Test\ImageProcessing\Photo;
 
 use App\Exceptions\ZipInvalidException;
 use function Safe\unlink;
